Allow Apache integration specify the server's trusted CA
Usually the host should be configured so it trusts its own Apache
Certificate Authority, even if it is self-created and not trusted
outside the host.

In case the user don't want to do it, two configuration options have
been provided to allow the user to specify the CA certificate or
the folder where the certificate is.